Socket.io events refactored into separate files. Quick cleanup of whitespacing and format.
