--- static/js/connect.js
@@ -1,4 +1,0 @@
-// //change as appropriate
-// var _Envs = {"development": false};
-
-// var realTalkUrl = _Envs.development ? "127.0.0.1" : 'http://realtalkclient.herokuapp.com';
